# apps/docs/src/openapi/jurnal-zitn-api.json
{
...
  "info": {
    "description": "HTTP API for Jurnal ZITN — the open-source trading journal.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Jurnal ZITN API",
    "version": "0.15.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/playbooks": {
      "get": {
        "responses": {
          "200": {
            "description": "{ items: Playbook[] }."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "List the caller's playbooks.",
        "tags": [
          "Playbooks"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "entryTrigger": {
                    "type": "string"
                  },
                  "exitCriteria": {
                    "type": "string"
                  },
                  "instrument": {
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "setupRules": {
                    "type": "string"
                  },
                  "timeframe": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created playbook."
          },
          "400": {
            "description": "VALIDATION_ERROR."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "Create a playbook.",
        "tags": [
          "Playbooks"
        ]
      }
    },
    "/api/playbooks/stats": {
      "get": {
        "description": "Groups the caller's latched flat trades by their soft playbook link and returns the same completed-trade statistics the performance surface uses. Trades whose playbook was deleted (or never had one) fold into a single entry with `playbookId: null`.\n",
        "responses": {
          "200": {
            "description": "{ items: PlaybookSetupStats[] }."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "Per-setup statistics, computed at read time.",
        "tags": [
          "Playbooks"
        ]
      }
    },
    "/api/playbooks/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Delete a playbook (soft link — trades are untouched).",
        "tags": [
          "Playbooks"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The playbook."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Get one playbook.",
        "tags": [
          "Playbooks"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated playbook."
          },
          "400": {
            "description": "VALIDATION_ERROR."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Update a playbook.",
        "tags": [
          "Playbooks"
        ]
      }
    },
...
    "/api/trade-plans": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "enum": [
                "pending",
                "executed",
                "missed",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items: TradePlan[] }."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "List the caller's pre-trade plans (optionally by status).",
        "tags": [
          "Trade Plans"
        ]
      },
      "post": {
        "responses": {
          "201": {
            "description": "The created plan."
          },
          "400": {
            "description": "VALIDATION_ERROR."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "Create a pre-trade plan (Pending).",
        "tags": [
          "Trade Plans"
        ]
      }
    },
    "/api/trade-plans/{id}": {
      "delete": {
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Delete a pre-trade plan.",
        "tags": [
          "Trade Plans"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The plan."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Get one pre-trade plan.",
        "tags": [
          "Trade Plans"
        ]
      },
      "patch": {
        "responses": {
          "200": {
            "description": "The updated plan."
          },
          "400": {
            "description": "VALIDATION_ERROR."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Update a pre-trade plan's fields.",
        "tags": [
          "Trade Plans"
        ]
      }
    },
    "/api/trade-plans/{id}/link": {
      "patch": {
        "description": "Body `{ positionId: uuid | null }`. The position must be the caller's and must not already be linked to another plan. Attaching a Pending plan marks it Executed; detaching an Executed plan returns it to Pending. PATCH (never GET) per the CSRF posture. Soft link — deleting the trade leaves the plan.\n",
        "responses": {
          "200": {
            "description": "The updated plan."
          },
          "400": {
            "description": "VALIDATION_ERROR — bad body or an already-linked trade."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND — no such plan or position."
          }
        },
        "summary": "Attach a plan to the trade it produced (1:1), or detach.",
        "tags": [
          "Trade Plans"
        ]
      }
    },
    "/api/trade-plans/{id}/status": {
      "patch": {
        "description": "Pending → executed / missed / cancelled. Repeating a terminal state is idempotent; moving between terminal states is a 400.",
        "responses": {
          "200": {
            "description": "The updated plan."
          },
          "400": {
            "description": "VALIDATION_ERROR — illegal transition."
          },
          "401": {
            "description": "No valid session."
          },
          "404": {
            "description": "NOT_FOUND."
          }
        },
        "summary": "Move a plan through its lifecycle.",
        "tags": [
          "Trade Plans"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}